cdl-mmkg-mvp/backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
motor>=3.3.0
sentence-transformers>=2.2.0
spacy>=3.7.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
PyPDF2>=3.0.0
python-docx>=1.1.0
aiofiles>=23.2.0
pymongo>=4.6.0